Updates FOGI reparameterization to get this feature working again, at least for 1Q.
Updates MatrixForwardSimulator so that FOGI reparameterization works again.  The ability
to include SPAM in the FOGI analysis required that we update the SPAM-computing parts
of the matrix forward simulator.  Additionally, some bugs were fixed in the FOGI SPAM
analysis itself, such as in the construction of the prep & povm gauge action matrices
(these seem to not have had any effect on the 1Q tests, but would for higher numbers of
qubits) where there was a typo related to indexing and a forgotten minus sign.  The model
parameter interposer was also updated (method `ops_params_dependent_on_model_params` added)
so that wrtfilter arguments could be properly accounted for in matrix fwdsim routines
which previously just punted with a NotImplementedError.