Adds StateSpace.sole_tensor_product_block_labels attribute.
This replaces many usages of statespace.tensor_product_block_labels(0)
where the labels of the first tensor product block are assumed to be
all there is to the state space.  This new attribute enfores this
assumption, throwing a ValueError if it isn't satisfied.  This commit
updates (hopefully) all the instances where .tensor_product_block_labels(0)
was used and .sole_tensor_product_block_labels should be used.

Incidentally, this commit also adds (temporarily) some debug statements to the
FOGI tools methods that are unrelated to the added attribute.