Fixes and updates FOGI analysis for LocalNoiseModels; particularly sslbl ordering issue.
Unlike the smq2Q models, a 2-qubit LocalNoiseModel by default contains
2-qubit gates in both directions, e.g., Gcnot:0:1 and Gcnot:1:0.  This
caused the discovery of a bug in that the sslbl ordering of Gcnot:1:0 was being
preserved, causing issues and incorrectness in the computed FOGI directions --
e.g. the intrinsic qtys of Gcnot:0:1 and Gcnot:1:0 were the *same* rather than
having their sslbls reversed.  This commit fixes this issue by:

1) adding sorting of sslbls to GlobalElementaryErrorgenLabel, facilitating
   correct comparison of the labels, e.g., that H(X,Y) on 0:1 is equal to
   H(Y,X) on 1:0.
2) adding sslbl mapping to EmbeddedOp errorgen_coefficient_array and errorgen_coefficients
   methods, since the embedded op's state space should NOT be assumed to be a
   meaningful subset of the EmbeddedOp.state_space labels.
3) Updates first_order_gauge_action_matrix and construct_fogi_quantities (in fogitools.py)
   to short-circuit embedding ops based on their sslbl ordering and not just space
   dimension.  Also adds debug statements related to ferreting out issue solved by next commit.