Fixes a bug in matrixtools.nullspace (forgotten .conjugate()!)
This bug was causing a failure within the construction of
relational FOGI directions within construct_fogi_quantities, where
an assertion failure precipitated a deep dive into what was going
wrong with the FOGI machinery.  As it turned out, the issue was that
the nullspace was being computed incorrectly for complex matrices,
which this commit fixes.