Updates setup_fogi within implicit (local noise) models to work with embedded ops.
Updates extract_std_target_mx function within setup_fogi to make use of
the recent updates to EmbeddedOp that properly map sslbls between the
EmbeddedOp and the op being embedded.  Also updates comparisons that
should have tested sslbls ordering but didn't.

As of this commit, the FOGI analysis of a LocalNoiseModel seems to
be working - or at least we've fixed the blatant errors now.