Switches nullspace -> nice_nullspace within gauge space intersection.
This leads to nice (if possible) vectors for the common gauge space
found between all the operations, and often results in the final FOGI
vectors & directions being simpler in terms of elementary errorgens.
This doesn't correct any bug per se, but it does make the results nicer.

In the 2Q FOGI test between explicit and cloud noise models, this update
causes the outputs to agree even after setting include_spam=True (in the
previous commit the outputs matched when SPAM was not included, but including
SPAM would cause the common gauge space to be a bit wacky and ultimately
different, more complex, FOGI directions to be selected).