Updates create_cloudnoise_circuits to work when max_idle_weight=0 (and fixes a bug).
A number of updates and a bug fix so that cloud noise circuits can be
constructed for a model without an idle gate, for which max_idle_weight=0.
The bug that was fixed was substituting 'max_idle_weight' with 'syntheticIdleWt'
when constructing the synthetic idle model (`sidle_model` in the code).

This commit also imports * from cloudcircuitconstruction.py in
pygsti/circuits/__init__.py so the contained routines are more accessible.