Adds a try/except guard to ConfidenceRegionFactory.compute_hessian for implicit models.
This try/except block allows model.num_nongauge_params to fail and fall back to
model.num_params, which will occur for implicit models.  This enables computation
of Hessians for implicit models, which we need for testing the Hessian spectrum of
models before and after FOGI reparameterization.  These tests are able to run for
LocalNoiseModel and CloudNoiseModel models now, though the spectrum isn't what I
expect so there's likely still some bugs in the FOGI analysis for implicit models
currently.