Begins adding API for extracting Kraus operators from pyGSTi operations.
Adds kraussop.py that contains interface (base) class.  Adds implementation
of this interface in DenseOperator and DenseUnitaryOperator, which required
adding a 'basis' argument to DenseOperator.__init__.  This update cascaded to
updating several of the DenseOperator subclasses, such as FullArbitraryOp.
'None' is allowed to be specified as the basis, but if so the Kraus op functionality
is unavailable.