Adds Kraus interface to StochasticNoiseOp and new representation classes.
The big update in this commit is the addition of new OpRepKraus
and OpRepRandomUnitary classes the the densitymx and densitymx_slow
evotypes, and OpRepRandomUnitary to the chp evotype.  The existing
OpRepStochastic class (only in these 3 evotypes) is updated to inherit
from OpRepRandomUnitary and has an updated signature, propagating to
updates in StochasticNoiseOp.