Adds 'only_global' implicit idle mode.
This new idle mode causes a detected global idle operation, e.g., an
'{idle}' gate or 'Gi' that acts on the entire state space, to be used
to simulate the empty layer but *not* other layers (unlike in the 'add_global'
mode).  This mode has been added for LocalNoiseModel and CloudNoiseModel,
and is useful for comparing to explicit models with an idle ('[]') gate.