Fixes a minor bug in cloud circuit construction to prevent a warning.
When an idle processor spec is internally constructed, the included
gate names should not contain gates that act on more qubits than are
in the processor spec, otherwise a warning will be generated about
unused model noise parameters.  This commit adds a filter to the gate
names used to construct the idle pspec, avoiding this warning.