Adds temporary debugging statements for debugging FOGI vector sparsity.
Adds several commented out lines in model.py that will be used when we
resume debugging the known issue that FOGI vectors need not respect the
same sparsity/locality as their corresponding FOGI directions.  This comes
about because the FOGI directions are not necessarily orthogonal on shared
subspaces (e.g. the Gx space for 2 relational quantities: one between Gx
and Gy and the other between Gx and Gz).  Perhaps we can perform the FOGI
direction construction in a way that preserves this desired property?  To
be picked up later.