Updates elementary errorgen basis from mxBasis to 'PP' in report tables.
Following the updates to elementary error generator construction (now
closely following the taxonomy paper), the elementary error generator basis
should usually be the un-normalized (traditional) Pauli-product basis when
the state space is composed of qubits.  This gives the intuitively expected
coefficients, e.g. 0.5pi for the Hamiltonian projection of a pi/2 rotation gate.
This commit updates the report quantities giving the projections onto gates and
error generators to use the 'PP' basis whenever the model space is entirely qubits,
and to only fall back on the model's superoperator basis in other cases.  This means
that the intuitive error generator coefficients and gate decomposition should occur
within reports for qubit models.  However, in other (e.g. qutrit model) cases, the
code still falls back to the model basis - and this may not be what we actually want.
FUTURE work: fix/improve the elementary errorgen basis selection for non-qubit models.