Fixes a bug in computation of POVM "map" used to compute POVM metrics (e.g., POVM fidelity)
Past updates have caused the Basis contained in a Model to be a TensorProductBasis
when there are more than two qubits, rather than a standard or explicit basis.  This
is fine, except within the computation of the POVM map we requested
basis.reverse_transform_matrix('std'), which gives a transform to the *equivalent*
'std' basis rather than *the* std basis of a given dimension.  The constructing code
assumed the latter, effectively assuming incorrectly that the model basis was a simple
(non-tensor-product) basis.  This commit fixes this issue by specifying the assumed
'std' basis explicitly.

Because this bug affected the POVM map construction, it affected all (and only) the POVM
metric values in the report, including the POVM infidelity, POVM trace distance, and
POVM diamond distance.