Adds ability to use statevec_slow evotype with WeakForwardSimulator
Implements a default outcome-for-shot method in WeakForwardSimulator
(though this is inefficient when the "shot" actually can compute a
probability distribution rather than a single outcome) that can be
used by the now-augmented 'statevec_slow' evotype.  The statevec_slow
evotype has been augmented with acton_random methods and new
OpRepRandomUnitary and OpRepStochastic classes to allow it to
work with WeakForwardSimulator even with models containing StochasticNoiseOp
operations.