Adds experimental code that is probably a dead end for making FOGI *vectors* local.
The goal is to use the freedom in constructing the FOGI dual "directions"
to choose a basis whose corresponding "primal" vectors share the same
locality as the directions.  This commit marks an (I think) failed attempt
at beginning this, but maybe some of it can be salvaged for later use.
This approach tries to contruct a valid and local primal basis alongside
the dual basis, and attempts to contrain the chosen FOGI directions by
ensuring orthogonal gauge-space directions.  This quickly gets complicated
when considering larger (3+ gate) relational sets, partially due to the
fact that the pseudoinverse of a block matrix cannot be performed
block-wise (particularly in our case where there's rank deficiency).

A more direct approach of omitting directions that lead to primal
vectors that aren't orthogonal to existing directions may work better?