A second, probably failed, attempt at finding local FOGI vectors in addition to directions.
This more direct approach constrains the created primal and dual vectors to be orthogonal
to the existing sets of dual and primal vectors, respectively.  This succeeds in giving
us a nice set of local FOGI directions and corresponding vectors, but fails to find
enough such vectors -- so this seems like too harsh of a constraint during construction.