Third failed attempt to produce local FOGI vectors.
This attempt tries to update the way relational quantities are constructed
so that it's easier to find orthogonal vectors to them.  It constructs FOGI
directions form the *separate* pseudoinverses of an existing set's elements,
which, unfortunately, doesn't work in general since the reference-frame (gauge)
intersection space will in general contain vectors that not contained in the
ranges of any/all of the constituent operation's gauge actions.

This commit leaves debug statements active and currently seems like a dead end.