Adds 'qibo' evotype and updates existing operator & evotype API to facilitate this addition.
The 'qibo' evotype uses the qibo package to perform circuit simulation,
and works, so far, with MapForwardSimulator.  A qibo state (rep) can be
either a state vector or std-basis density matrix, and module-level variables
in pygsti/evotypes/qibo can control which mode is used, along with the
number of shots used for internal-to-qibo weak forward simulation.  Holding
a 'std' basis density matrix required that dense states be given a 'basis'
argument, and that several other op & effect rep types and operations also
take 'basis' arguments:
 - dense superop reps (only needed for implementing to_dense though)
 - (new) kraus reps
 - dense state reps
 - dense op objects (because of rep)
 - dense state objects (because of rep)

Other operations/states/effects that would sometimes construct
dense ops (ComposedOp, EmbeddedOp, ExpErrorgenOp, LindbladErrorgen,
ComplementPOVMEffect, ComposedPOVMEffect, ConjugatedStatePOVMEffect,
more?) and aren't given a 'basis' argument currently punt or are
hacked, often resulting in the basis being set to None.  This is
marked for future fixing with "UNSPECIFED BASIS" comments.  It seems
best to create a standardized .basis attribute of state and maybe
op & effect reps so that these operations can access the basis (or
lack thereof - it should be allowed to be None) of a contained
operator.

Overall, this commit includes a partial plumbing of 'basis' arguments
- particularly for those representation and operation classes that
take dense vectors or matrices - so that the evotype has access to
basis information.  With a standard .basis attribute this infomation
will be spread and accessible in all the different representation
classes, which seems ideal, given that other 3rd party simulators
may also benefit from this information. Adding a .basis attribute shouldn't
add much/any performance overhead since the basis is just passed
around by reference.

At this point, the qibo evotype seems to work for simple 1Q
circuits with the 'map' simulator, but more testing is needed.