Adds IdentityPlusErrorgenOp as an alternative to ExpErrorgenOp.
The new IdentityPlusErrorgenOp class implements the first-order approximation
to exp(L), 1+L, as an alternate way of using an error generator within an
operation.  This has the advantage that no matrix exponentiation is required,
potentially saving computational time (and number of paths), though 1+L is only
CPTP for small L, and so may lead to Kraus-decomp errors within evotypes like
qibo.  Additionally, the first order approximation doesn't compose as nicely
nor does a low-weight L capture all the expected "physical" errors from such a
process - so there are plusses and minuses to it.  In any case, it's now an
option to experiment with, and can be accessed by specifying '1+(<LindbladType>)'
parameterizations, e.g. '1+(CPTPLND)' or '1+(H+S)'.

This commit also adds support for more explicit parameter/gate-type naming,
updating 'CPTP' to 'CPTPLND' when used as a Lindbladian param type (this is
more accurate) and adding the ability to specify exponentiated errorgenerator
paramterizations explicitly with 'exp(.)' names, e.g. 'exp(CPTPLND)' or 'exp(H+S)'
instead of just 'CPTPLND' or 'H+S' (though these default to the exp versions).

Specific support for IdentityPlusErrorgenOp is provided by the new OpRepIdentityPlusErrorgen
representation class.  This commit only implements this for the densitymx_slow
evotype, though it should be straightforward within other evotypes - and *easier* /
*simpler* to implement than the OpRepExpErrorgen class (e.g. could do this for qibo).
If this new rep type is unavailable, IdentityPlusErrorgenOp falls back to using a
dense op.

Testing has been minimal at this point - a 1Q GST example with densitymx_slow.