Add caching for circuit PTMs and deduping and identity removal for the available fiducials
This commit adds basic caching support to avoid re-calculating the PTMs for the available fiducial list when generating the prep cache and measurement cache, which typically use the same set of available circuits and as such the same PTMs. Also add a function for removing effective identities and circuits with duplicate PTMs from the available fiducial list to decrease the search space.