Even more changes to the caching and deduping routines
Biggest changes are to the deduping routine which is now more efficient because we now winnow the list of circuits to search through for duplicates on the fly, reducing the number of pairwise checks along the way, big speedboost from this. Also switched to using the frobenious norm to identify duplicate PTMs as this was much faster in practice than np.allclose.  Also updates the cache to use the string reps of circuits as keys instead of the circuits themselves to reduce the cost of repeatedly hashing circuits which can be more expensive than straight strings, modest performance enhancement from this