Another round of Kivy GUI additions.  Data explorer app is beginning to take shape.
Still very rough and experimental, but contains proof-of-concept implementations
for a number of the important ingredients for moving forward with a Kivy-based GUI:
- color box plots (scatter plots will be similar)
- color histogram w/line
- colormaps into figures (some finished, some not)
- hover labels into figures (just proof of concept in ColorBoxPlot)
- latex rendering (via SVG - somewhat slow but working)
- figure sizing (basic idea is there, but needs lots of tweaking/fixing)

Note that at this commit the color box plot uses texture coordinates
to color boxes, whereas the color histogram uses custom vertex and
fragment shaders.  The latter approach is better (uses less video mem)
and preferred -- the texture-based approach was an initial attempt
before figuring out how to do it better.  So we should update the
color box plot code in the future to use the better approach.