Adds backend-frontend division to Kivy-based data explorer and other minor updates.
Continues progress on the prototype data explorer Kivy app, adding the
key ability to use an ipython kernel to perform all the object building
and manipulation while displaying the results in the Kivy-powered front
end.  Data is tranferred via the same ZeroMQ socket interface used by
jupyter notebooks, and currently amounts to sending json'd from kernel
to client.  A "mock" kernel is also added for debugging purposes.  The
same basic GUI functionality that existed before the front/back division
exists at this point, which a few minor additions (e.g. a delete button
for deleting library items and a comment field for figures).  There's
still lots of work to do, but at this point the app can and was demoed.
Here's a list of some of the biggest ideas and plans for future features:

1) it would be useful to have a way to export plots and/or whole reports
  to PNG or PDF files.  Maybe we can use the latex/html rendering, but this
  is rusty and would take some time.  The Kivy export_to_png function
  seems useful as a minimal export capability.

2) the backend implementation is still very clunky, and lacks a usable
  API.  Ideally we could import the backend code more smoothly, and
  provide an API for navigating and accessing information on the back
  end.

3) we need to work out a way to duplicate a report and tabs within a
  report so that after a single report/tab/figure is make others can
  be made more quickly.  At the figure level this is basically done,
  as figures are pretty easy to create.  At the report level, we could
  consider simply loading/replacing the library with one that has similar
  names, then re-loading all the figures (?), in which case it would be
  nice to be able to construct libraries entirely within the backend
  python interface.  At the tab or group-of-figures level, we could
  create "template" library entries that can be used as selector values
  that point to a given library item and that can be switched in a way
  that triggers re-generation of all "connected" figures (e.g. a 'modelA'
  template entry that can be switched between all or several of the
  available models, similarly to the dropdown in HTML reports).  Perhaps
  there could be a way to create and save your own presets?

4) Expanding the types and managing of annotations would be helpful
  for "marking up" a report.  Adding descriptions of figures and
  adding an ability to add headings and notes on the screen seems
  valuable.