Updates internal format of the compact EVD cache
I was originally returning both the components of the rank decomposition and the eigenvalues for the cache, but I never actually used the eigenvalues so I am dropping them from this structure. If I think of an application where I need to know what these are I can always reverse these changes very easily.