Penalty term based on number of instances of specified gates
Adds an additional penalty term to the objective function where in a penalty is added to the score for each instance of a particular gate. This is passed in by the user in the form of a dictionary where the keys are gates and the values are the penalty weights.