Fixes a bug in wildcard budget processing.
This bug caused an assertion error to be raised (when it shouldn't be)
because tolerances were mismatched (1e-6 was used in one place and 1e-8
used in another).  This caused too little processing to be done
for a given circuit which was then picked up by a downstream (lower tolerance
and more stringent) check.  This should (hopefully!) fix the unit test
failure we started getting, seemingly randomly, in recent commits.