Following up on the last commit, adds ValueError back to 2 system tests.
In the last commit we replaced ValueError with LinAlgError in two places,
since that's what SciPy was raising.  This wasn't quite true - *some* versions
of Scipy's expm function raise a LinAlgError when the exponentiated matrix isn't
square while others raise a ValueError.  So the system tests need to check for
*both* types of exceptions.  This issue also indicates that it probably wasn't
any pyGSTi update that broke these IO system tests in the first place - I think
it was just an updated version of SciPy.  Should be fixed now, with system tests
asserting that either type of exception is raised.