Adds unit tests for qibo evotype and adds minor bug fixes and tweaks.
"Tweaks" here are mainly for getting 1+(CPTPLND) parameterization to
work with qibo evotype.  Note that OpRepIdentityPlusErrorgen exists
in the densitymx_slow but *not* in the densitymx evotype.  (This would
be trivial to add, but it's unclear whether this oprep type will be of
much use, so we haven't bothered mirroring it yet.)