Add more handles onto the eigenvalue tolerance thresholds
Adds additional handles for setting and piping in a use specified value of the minimum eigenvalue threshold tolerance. I think the Cholesky decompositions we were seeing fail were effectively due to the Jacobian whose psuedoinverse we were updating being poorly conditioned due to an overly generous (i.e. too small) minimum eigenvalue tolerance. This was technically a variable that could be set but didn't have a way to do so at the top level and the hardcoded value resulted in issues.