Fixes bugs from feature-qibo branch merge caught by system tests.
Re-creates reference models (json'd file that break when signature of
StateRep objects changes, e.g. by addition of basis arg), fixes an argument
bug, and loosens to_dense tolerance for CPTPLND map -> to_dense conversion
from 1e-11 to 3e-10.  The need for this last change is peculiar and I'm not
sure why any of the qibo updates would require this.  We could look into this
more later, but for now we should just keep an eye on this and see if any other
more egregious to_dense failures occur that are clearly the sign of a bug and
not just numerical/roundoff issues.