Adds catching of more generic ArpackError to svd exception processing in matrixtools.independent_columns
In a very recent commit, we updated the caught ArpackNoConvergence exception's location
within scipy, thinking that this was a simple name update.  It seems like scipy.sparse's
SVD is now failing with the more generic ArpackError exception when previously it failed
with ArpackNoConvergence, so we're updating the exception block to catch both types of
exceptions now.  I *think* this is all fine and there's no issue here, but we should keep an
eye on this as the new exception handling may also mask other more important SVD failures.