Update implementation of inverse trace for edge case
This adds a check for what was previously an edge case, but now is a common occurrence when applying this to FPR, where the update doesn't change the rank of the base matrix because that matrix is full rank already. In this case we fall back to the standard Woodbury inverse identity.