Implement new greedy search algorithm
This commit introduced a new greedy search algorithm leveraging the same low-rank update tricks that we use in other parts of the experiment design codebase. This also introduces a new deduping routine specialized to the case where our gate set only generates the clifford group in which we leverage the fact that the PTMs of clifford operations are signed permutation matrices. This coupled with hash table magic (using python dictionaries which use them internally) gives an amortized O(N) deduping routine in this case.