Initial, mostly complete, implementation of MongoDB serialization.
Adds ability to serialize pyGSTi objects to a MongoDB database by
creating routines parallel to those used for disk (filesystem) I/O
that read and write MongoDB documents.  This commit adds support for
many of pyGSTi's objects, and main types (experiment designs, datasets,
results, protocols) have been tested to be able to be read and written
to a MongoDB instance.

Most of the implementation lies within the new io/mongodb.py module,
which parallels metadir.py functionality for a database.  Issues arise
in that the paths used for filesystem I/O serve as both an ID (e.g.,
for a DB record) and a hierarchical link, indicating the parent and children
of a record.  The current MongoDB implementation uses subcollections
to store related data, but this still seems a bit clunky and can result
in many "nested" MongoDB subcollections.  Also, removal of data is less
straightforward in the DB setting.

Overall, in addition to basic bug fixing, we still need to look into
the following:
- ability to write over (& update) existing DB documents
- ability to remove documents cleanly.
- resolve issue of subcollection nesting.