Adds more meaningful error message when StandardGST.run fails needing a target model.
Usually a StandardGST protocol either needs to be given a target model or
is run(...) with an experiment design that can produce one via a processor
spec.  This commit adds better error handling and reporting for the
case when a target model is not given and can't be obtained via the edesign.