Fixes serialization bug in QuditProcessorSpec and updates map_qudit_labels.
Removes a spurious reference to "_symplectic_reps" from QuditProcessorSpec
(this is a member of QubitProcessorSpec but not QuditProcessorSpec), along with
a repeated argument in from_nice_serialization's call to QuditProcessorSpec.__init__,
causing QuditProcessorSpec to error when un-serialized.

Also updates QuditProcessorSpec.map_qudit_labels so that a QubitProcessorSpec
object is returned when `self` is a QubitProcessorSpec.  This seems natural
and expected, and I think if we don't mind a small hit to efficiency we could
just have QubitProcessorSpec.map_qubit_labels fall back to this function now.