Allow reduced fiducial sets for L=1
Modifies the LSGST creation routine to allow for a reduced set of fiducials to be used for L=1 when include_lgst is set to false. The scenario this is meant to enable is one where at L=1 a minimally IC set of fiducials is used in conjunction with FPR with respect to a larger set of fiducials at larger circuit depths.