Adds ability (or fix?) to be able to serialize results containing wildcard budgets.
Updates WildcardBudget to inherit from NicelySerializable and updates what
is stored within GST Estimate objects to be a *serialized* WildcardBudget
rather than the object itself.  This allows the Estimate to serialize
itself without any modification, using the standard metadir framework.
On the other hand, it necessitates updating the report construction
(in factory.py) to initialize a WildcardBudget object from the Estimate's
serialization, which is not performed.

This update may therefor break user code that assumes that
estimate.extra_parameters['unmodeled'] is a WildcardBudget.