Updates fiducial pair reduction to only look for pairs that "access" model parameters.
In the past, FPR has sought a set of fiducial pairs that access all of the
degrees of freedom amplified by each germ, namely the degrees of freedom of
the commutant of each germ.  This assumed that the model parameters spanned
the space of all these commutants, which isn't always true.

This small commit adds a minimal update to the FPR routine that tries to
account for the space of model parameters when selecting fiducial pairs.
It hasn't been tested much yet, and doesn't address the issue of selecting
multiple fiducial pairs for different germs that access the same model
paramter(s).

to the model parameters