Adds support for MongoDB serialization of confidence region factories.
Adds necessary read/write from/to MongoDB methods to CachedObjectiveFunction
(forgotten previously).  Adds MongoSerializable class as an extension of
NicelySerializable that gives the object being serialized access to a
MongoDB database for placing large/complex components.  This framework
is needed by ConfidenceRegionFactory, which uses gridfs to store Hessian
and Hessian projection data when serialized to a MongoDB database.  This
is necessary because (currently) the maximum MongoDB document size is 16MB.

This commit also fixes a few minor bugs brought out by testing the
serialization of large & nested trees of results, data, and edesigns
that included confidence region factories and wildcard budgets.