Fixes bug in StandardGSTDesign.copy_with_maxlengths where nesting flag is reset.
A call to a generic truncation function within copy_with_maxlengths reset
the 'nested' flag (appropriately I think, since the edesign may no longer
be nested) and copy_with_maxlengths should have set the flag so it conveyed.
This commit updates the function so it does convey the nested flag and
copy_with_maxlengths works as it should.