Generalizes the "diamond distance wildcard model" to allow for other metrics.
Renames 'PrimitiveOpsDiamondDistanceWildcardBudget' to
'PrimitiveOpsSingleScaleWildcardBudget' and also refactor the
base PrimitiveOpsWildcardBudget class so that the new wildcard budget
has just a single parameter (alpha) while still retaining the useful
machinery of the base class.

Updates 'ddist_wildcard' to 'wildcard1d' as a bad-fit option when
running GST, and adds a 'wildcard1d_reference' attribute to GSTBadFitOptions.

By default, badfit_options.wildcard1d_reference == 'diamond distance' and
the function _compute_1d_reference_values_and_name computes the diamond
distance as before.  If we want to add other "built-in" options this just
amounts to adding cases to this function.  Allowing custom user functions
could also be done by allowing an object in addition to string values, but
let's leave this as future work if it's ever needed.

This commit also moves some of the wildcard optimization code from gst.py
into wildcardopt.py, which seems more appropriate, and refactors the wildcard
optimization code a bit to lessen the amount of duplicated code.

The wildcard bar plot in the report remains diamond-distance specific, and
is only triggered when the unmodeled error is a PrimitiveOpsSingleScaleWildcardBudget
AND its .reference_name == 'diamond distance'.  So, if we want to add additional
metric options and report them we'd need to tweak the DiamondDistanceWildcard code in
factory.py and figure generation code in section/goodness.py and Goodness_unmodeled.html
files where "ddist" and "Diamond Distance" are still hardcoded.  Note, however,
that we'd also need to change the figure caption & title.

The plot class itself, now named WildcardSingleScaleBarPlot, has been generalized
to accept a user-specified title for the reference values along with a
PrimitiveOpsSingleScaleWildcardBudget.