Refactor deduping and final testing code
Refactors the deduping code to combine the two separate functions for deduping the candidate fiducial list into a single one that branches on the case where we are assuming a Clifford gate set. Also refactors a printing subroutine to reduce amount of repeating. Finally adds some docstring and debug clean-ups.