Adds forgotten sqrt(2) factors to Gxx and Gzz unitary matrices in internalgates.py
This would cause Gzz and Gxx gates to be created incorrectly, e.g., in a processor
spec whenever they were used.  Note that this bug doesn't affect the smq2Q_XYXX and
smq2Q_XYZZ modelpacks which use a (correct) hard-coded unitary in the model creation
routines in modelconstruction.py.