Updates find_germs_breadthfirst_greedy to at least give a better error msg.
Updates this function to avoid a control path that can occur where a
variable is left undefined.  I ran into this when running germ selection
with no (zero) forced germs - at least I think this was the salient aspect
that caused the error.  I was thinking this update would fix the error but
alas no.  We'll postpone tracking this down right now as other methods work
fine, but we should come back to this later.