Updates MongoDB serialization so that overwrite=False behaves in a "smart" way.
Previously, setting overwrite=False when saving an object to a mongoDB
would fail if a document with the destination id already existed.  With
this update, an error is *not* raised in this case if the existing mongoDB
document matches the one being written.

This behaviour is advantageous when dealing with nested objects, and
paves the way toward a serialization that doesn't create duplicate
documents for objects that are contained within other objects.  For
example, saving a ProtoclData object that contains a reference to an
already-saved experiment design might be able to avoid re-writing that
experiment design by checking that a matching document exists in
the database already.  This also fixes an even more blatent issue that
when an object with standalone sub-objects was saved multiple times with
the same ID this would continue to zombie the old and create new standalone
object documents.

The use of a new function, pygsti.io.mongodb.prepare_doc_for_existing_doc_check,
massages a document that is being saved to smooth out aspects of a
strict dictionary comparison with a database document.