Fixes ability to mongodb-serialize tree nodes with children.
Previously, if edesign had a different MongoDB ID than the data or
results object that contained it, serialization would fail because
the data or results object would expect to be able to query for the
edesign object using the same id.  Now a `preloaded_edesign` can be
passed to TreeNode._init_children(...) explicitly.