Revamps MongoDB serialization format, making it less filesystem analogous.
The initial MongoDB serialization support basically duplicated the
JSON-based filesystem serialization to write MongoDB documents.  This
update restructures the MongoDB document format used by pyGSTi objects,
while keeping the outward API nearly the same.

From an external perspective, the main change is that experiment design,
data, and results objects are no longer given path-like ids but will have
normal database identifiers.  These IDs are generated the first time an
object is written to the database (and returned by the "write" method)
and are held within the Python object for future writes.  You no longer
write *to* a path/ID, you simply write an object: either a new ID
is generated or an existing stored ID is used as appropriate.

Within the database, these IDs are used to link the various documents,
and now a parent object will hold the IDs of it's children/attribute
documents within its "main" document.  This results in a more typical
data model and allows documents to serve in a many-to-one capacity,
improving space efficiency.  For exampmle, an experiment design can
be linked to multiple protocol data objects and a circuit document
can be referenced by multiple experiment designs and/or data sets.

Internally, this commit makes MongoSerializable, which used to be
a separate and somewhat ad-hoc base class, a base class of
NicelySerializable (the base class for objects that have simple
JSON serializations).  A MongoSerializable object is one that can
be serialized to a MongoDB database, and so this update makes everything
that is JSON-serializable (most pyGSTi objects) also MongoDB-serializable,
which is nice and cohesive.  Furthermore, even pyGSTi's large
non-NicelySerializable objects (edesigns, protocol data, results,
protocols, datasets, estimates, confidence region factories,
cached objective fns) *are* MongoSerializable, and so are brought
onto the same footing as most of the rest of pyGSTi's objects.

Object removal is more difficult under the new data model, since
objects don't have clear owners and so it's difficult to tell
whether the members/attributes of an object should always be removed
along with it.  The `RecursiveRemovalSpecification` object is meant
to give some control over this, but users should be very careful
with removing database documents.

All in all, this commit represents a significant upgrade to the
MongoDB serialization capabilities in pyGSTi, but should be tested
more before we rely on it.