Update CI workflows
Update the OSes and the python versions used for unit-testing/CI on github actions. Adds MacOS11 Big Sur to the list of operating systems tested and updates the ubuntu runner to 20.04 (18.04 is deprecated on github actions and will go away eventually). Also increments the minimum and maximum python versions tested on up by one, so now we're testing versions 3.7-3.10.