Second (hopefully final) attempt to fix "final iteration estimate" parameterization.
Moves the 'convert_model_to' argument processing from gaugeopt_custom
to gaugeopt_to_target in order to fix a subtle bug from adding a
'model = model.copy()' line.  The subtley came from the fact that
a reference to the model passed into gaugeopt_custom was used via
function closure within the gauge optimization objective function,
and so adding the .copy() broke this connection and caused the model
actually used in gauge optimization to *not* be converted anymore (!).
This commit moves all this conversion hoopla before we construct the
gaugeopt objective function so things work as intended.